+++ app/controllers/editor/datasets_controller.rb
@@ -127,7 +127,8 @@ class Editor::DatasetsController < ApplicationController
       :organization_id, :name, :description, :slug, :logo, :logo_cache, :released,
       :git_repository, :release_date, :info_what, :info_who,
       :info_when, :info_funded_by, :info_citation, :subjects,
-      :age_min, :age_max, :time_frame, :polysomnography, :actigraphy, :doi
+      :age_min, :age_max, :time_frame, :polysomnography, :actigraphy, :doi,
+      :featured
     )
   end