+++ db/migrate/20200404224211_add_featured_to_datasets.rb
@@ -1,0 +1,6 @@
+class AddFeaturedToDatasets < ActiveRecord::Migration[6.0]
+  def change
+    add_column :datasets, :featured, :boolean, default: false, null: false
+    add_index :datasets, :featured
+  end
+end